ci: Clean up PyPI and dockerhub releases
Now releases are done via tags:

- `service-v*` tags release the service container to dockerhub
- `client-v*` tags release the client package to PyPI

Both workflows validate if the tag version matched the version from the
repository. Version of the service is distinct from the version of the
client.

Removed the unnecessery steps from test workflow.

Dependencies for the Pylon service are now optional (only client
dependencies are required)

Issue: BACT-230